+++ copy_with_extension_gen/test/generated_code_test_cases/test_cases_exceptions.dart
@@ -48,3 +48,20 @@ class TestNullability {
   // If a field is nullable, you can change the type of the constructor parameter to be non-nullable. However, if you do this, an exception may be thrown because the constructor only accepts non-nullable parameters. The issue is that we cannot guarantee that the parameter will be non-null when calling the constructor in the `copyWith` function. Therefore, even if the constructor constructs an object with a nullable field, it is currently impossible to achieve this in the implementation.
   final int? nullableWithNonNullableConstructor;
 }
+
+@CopyWith(copyWithNull: true)
+class CopyNullSuper {
+  const CopyNullSuper({this.superField});
+
+  final int? superField;
+}
+
+@ShouldThrow(
+  'Class "CopyNullSubclass" has nullable fields and extends a class annotated with `@CopyWith(copyWithNull: true)`, but does not enable `copyWithNull` itself. Annotation options are not inherited. Add `@CopyWith(copyWithNull: true)` to this class, or enable `copy_with_null` globally in `build.yaml`.',
+)
+@CopyWith()
+class CopyNullSubclass extends CopyNullSuper {
+  const CopyNullSubclass({super.superField, this.subField});
+
+  final String? subField;
+}